requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-WTF
WTForms
email-validator
python-dotenv
psycopg2-binary
requests
gunicorn
alembic
openpyxl

